+++ xdef/changelog.md
@@ -1,8 +1,10 @@
 # Version ${version}, release-date ${release.date}
 
-# Version 42.2.28, release-date 2025-06-08
+# Version 42.2.28, release-date 2025-06-10
 * If the `xdef_string_codes` property is set (org.xdef.XDConstants.XDPROPERTY_STRING_CODES), for items,
   where there is no validation method declaration, the character dataset is not checked.
+* Fixed a bug when translating source X-definitions where XML syntax errors would throw an exception with
+  the message "Internal error", but no information about the cause.
 
 # Version 42.2.27, release-date 2025-05-22
 * Fixed incorrect XDEF823 error reporting in values from parsing data in attributes or text nodes that are